rag-service/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
langchain
langchain-openai
langchain-chroma
langchain-community
pypdf
chromadb